requirements.txt
python-dotenv==0.14.0
Django==3.2.7
djangorestframework==3.13.1
psycopg2-binary==2.9.5
celery==5.2.7
redis==4.3.3
requests==2.28.1